Added an apps.py submodule to load the appconfig as recommended by django
This is why webhook registrations were not getting picked up correctly by tox (pytest) in djangomain
